fix: address claude hubspot review
